# Simplify `UTxOIndex`.