## Issue

ADP-3061

## Summary

This PR removes the use of `NonEmptySet` from `UTxOIndex`, and redefines it in terms of [`MonoidMap`](https://hackage.haskell.org/package/monoidmap/docs/Data-MonoidMap.html) and `Set`:
```patch
  data UTxOIndex u = UTxOIndex
      { indexAll
-         :: !(      Map Asset (NonEmptySet u))
+         :: !(MonoidMap Asset         (Set u))
      , indexSingletons
-         :: !(      Map Asset (NonEmptySet u))
+         :: !(MonoidMap Asset         (Set u))
      , indexPairs
-         :: !(      Map Asset (NonEmptySet u))
+         :: !(MonoidMap Asset         (Set u))
```

This enables us to simplify several functions. For example:
```patch
- insertEntry m a =
-     Map.alter (maybe (Just createNew) (Just . updateOld)) a m
-   where
-     createNew = NonEmptySet.singleton u
-     updateOld = NonEmptySet.insert u
+ insertEntry m a = MonoidMap.adjust (Set.insert u) a m
```

It also allows us to remove one of our dependencies on the `strict-non-empty-containers` library. (We eventually remove this in https://github.com/input-output-hk/cardano-wallet/pull/3984.)

## Performance

A very slight performance improvement was observed over 10 consecutive runs of the `utxo-index` benchmark compiled with GHC `8.10.7` and an optimisation level of `-O1`:
| Benchmark<br/>`utxo-index`| Mean total run time<br/>over 10 runs (3SF) |
| :-- | :-- |
| `master` | `11.0` seconds |
| After applying this PR | `10.7` seconds |

## Sample benchmark runs

### Before applying this PR
```
Benchmark utxo-index: RUNNING...
All
  Constructing a UTxO index from a map with ada-only bundles
    with 1,000 entries:   OK (1.39s)
      1.07 ms ±  39 μs
    with 10,000 entries:  OK (0.59s)
      14.3 ms ± 1.2 ms
    with 100,000 entries: OK (0.70s)
      171  ms ± 7.9 ms
  Constructing a UTxO index from a map with (ada, NFT) bundles
    with 1,000 entries:   OK (0.80s)
      3.51 ms ± 277 μs
    with 10,000 entries:  OK (0.34s)
      51.5 ms ± 3.4 ms
    with 100,000 entries: OK (2.26s)
      682  ms ± 5.3 ms
  Constructing a UTxO index from a map with diverse bundles
    with 1,000 entries:   OK (0.61s)
      4.54 ms ± 391 μs
    with 10,000 entries:  OK (0.70s)
      71.0 ms ± 4.2 ms
    with 100,000 entries: OK (3.56s)
      1.116 s ±  42 ms
```

### After applying this PR
```
Benchmark utxo-index: RUNNING...
All
  Constructing a UTxO index from a map with ada-only bundles
    with 1,000 entries:   OK (1.02s)
      1.02 ms ±  86 μs
    with 10,000 entries:  OK (0.54s)
      12.9 ms ± 768 μs
    with 100,000 entries: OK (0.68s)
      167  ms ± 7.2 ms
  Constructing a UTxO index from a map with (ada, NFT) bundles
    with 1,000 entries:   OK (0.76s)
      3.44 ms ± 195 μs
    with 10,000 entries:  OK (0.33s)
      51.0 ms ± 3.7 ms
    with 100,000 entries: OK (2.24s)
      674  ms ± 9.7 ms
  Constructing a UTxO index from a map with diverse bundles
    with 1,000 entries:   OK (0.74s)
      4.41 ms ± 204 μs
    with 10,000 entries:  OK (0.40s)
      69.4 ms ± 5.7 ms
    with 100,000 entries: OK (3.33s)
      1.072 s ±  49 ms
```