@jonathanknowles 2023-07-04
I have manually verified that this PR does not cause any failures in the following test suites:
- `cardano-wallet-primitive:test:test`
- `cardano-wallet:test:unit`
- `cardano-wallet:test:integration`